+++ metrics-util/src/lib.rs
@@ -2,25 +2,20 @@
 #![deny(missing_docs)]
 #![cfg_attr(docsrs, feature(doc_cfg), deny(rustdoc::broken_intra_doc_links))]
 
-#[cfg(feature = "handles")]
-mod bucket;
-#[cfg(feature = "handles")]
-#[cfg_attr(docsrs, doc(cfg(feature = "handles")))]
-pub use bucket::AtomicBucket;
-
 #[cfg(feature = "debugging")]
 #[cfg_attr(docsrs, doc(cfg(feature = "debugging")))]
 pub mod debugging;
 
-#[cfg(feature = "handles")]
-mod handles;
-
 mod quantile;
-pub use quantile::{parse_quantiles, Quantile};
+pub use self::quantile::{parse_quantiles, Quantile};
 
 #[cfg(feature = "registry")]
 #[cfg_attr(docsrs, doc(cfg(feature = "registry")))]
 pub mod registry;
+
+#[cfg(feature = "storage")]
+#[cfg_attr(docsrs, doc(cfg(feature = "storage")))]
+pub mod storage;
 
 mod common;
 pub use common::*;
@@ -31,17 +26,8 @@ pub use key::CompositeKey;
 mod kind;
 pub use kind::{MetricKind, MetricKindMask};
 
-mod histogram;
-pub use histogram::Histogram;
-
 mod recoverable;
 pub use recoverable::RecoverableRecorder;
-
-#[cfg(feature = "summary")]
-mod summary;
-#[cfg(feature = "summary")]
-#[cfg_attr(docsrs, doc(cfg(feature = "summary")))]
-pub use summary::Summary;
 
 pub mod layers;